Finish part 2 of the puzzle

+++ include/bank_processor.h
@@ -1,7 +1,8 @@
 #ifndef BANK_PROCESSOR_H
 #define BANK_PROCESSOR_H
 
+#include <stdint.h>
 #include <stdio.h>
-int process_banks_from_file(FILE *file);
+uintmax_t process_banks_from_file(FILE *file);
 
 #endif // !BANK_PROCESSOR_H
