+++ include/joltage_calc.h
@@ -1,6 +1,7 @@
 #ifndef JOLTAGE_CALC_H
 #define JOLTAGE_CALC_H
 
-int get_joltage_from_gate(char *gate);
+#include <stdint.h>
+uintmax_t get_joltage_from_gate(char *gate);
 
 #endif // !JOLTAGE_CALC_H
